Removed dependency on VS Telemetry.
