+++ src/vstest.console/vstest.console.csproj
@@ -60,10 +60,5 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup Condition="'$(TargetFramework)' == 'net451'">
-    <PackageReference Include="Microsoft.VisualStudio.Telemetry">
-      <Version>15.3.799-masterDDDBA9E4</Version>
-    </PackageReference>
-  </ItemGroup>
   <Import Project="$(TestPlatformRoot)scripts\build\TestPlatform.targets" />
 </Project>
